+++ SPTracer/src/Model/Camera.h
@@ -12,12 +12,12 @@ namespace SPTracer
 		Vec3 p;				// center of projection
 		Vec3 n;				// image plane normal
 		Vec3 up;			// up direction
-		float f;			// distance to image plane
-		float iw;			// image width
-		float ih;			// image height
-		float icx;			// image center x
-		float icy;			// image center y
-		float t;			// time of exposure
+		double f;			// distance to image plane
+		double iw;			// image width
+		double ih;			// image height
+		double icx;			// image center x
+		double icy;			// image center y
+		double t;			// time of exposure
 	};
 
 }
