+++ SPTracer/src/Window.cpp
@@ -110,14 +110,14 @@ std::shared_ptr<Window::ImageUpdater> Window::GetImageUpdater()
 
 std::string Window::NewWindowClassName()
 {
-	static const int len = 16;
+	static const size_t len = 16;
 	static const char alphanum[] =
 		"0123456789"
 		"ABCDEFGHIJKLMNOPQRSTUVWXYZ"
 		"abcdefghijklmnopqrstuvwxyz";
 
 	std::ostringstream ss;
-	for (int i = 0; i < len; ++i) {
+	for (size_t i = 0; i < len; ++i) {
 		ss << alphanum[std::rand() % (sizeof(alphanum) - 1)];
 	}
